Make this work with Next.js 12.1.x
With 12.1.x you receive a following error otherwise:

```
Invariant Violation: Could not find "client" in the context or passed in as an option. Wrap the root component in an <ApolloProvider>, or pass an ApolloClient instance in via options.
```

Seems like _app is NOT processed during the server render for some reason when _document is present.